jpa-15/7

+++ src/main/java/com/javabykhang/project2pf/repository/RentareaRepository.java
@@ -5,5 +5,5 @@ import com.javabykhang.project2pf.repository.entity.RentareaEntity;
 import java.util.List;
 
 public interface RentareaRepository {
-    List<RentareaEntity> findValueByBuildingID(int id);
+    List<RentareaEntity> findValueByBuildingID(Long id);
 }
